requirements.txt
groq
python-dotenv
SpeechRecognition
pyttsx3
pyaudio
PyPDF2
opencv-python
numpy